restricted delete user for super admin

+++ app/datatables/user_datatable.rb
@@ -30,7 +30,9 @@ class UserDatatable < ApplicationDatatable
         action: content_tag(:div, class: 'btn-group') do
           concat(link_to(fa_icon('eye padding-right'), resource_path(record)))
           concat(link_to(fa_icon('pencil padding-right'), edit_user_path(record)))
-          concat(link_to(fa_icon('trash-o padding-right'), resource_path(record), method: :delete, data: {confirm_swal: 'Are you sure?'}))
+          unless record.user_type_name == 'Administrator'
+            concat(link_to(fa_icon('trash-o padding-right'), resource_path(record), method: :delete, data: {confirm_swal: 'Are you sure?'}))
+          end
         end
       }
     end
